@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
-import { AuthService } from "./services/api/auth-service/auth.service";
+import { AuthService } from './services/api/auth-service/auth.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
   constructor(public authService: AuthService) { }
