@@ -1,15 +1,14 @@
 import { Injectable } from '@angular/core';
-import { AbstractApiService } from "../api.service";
-import { Assignment } from "./assignment";
-import {HttpClient, HttpHeaders} from "@angular/common/http";
-import {Observable, of, pipe, shareReplay} from "rxjs";
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Observable, of, shareReplay } from 'rxjs';
+import { AbstractApiService } from '../api.service';
+import { Assignment } from './assignment';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AssignmentService extends AbstractApiService<Assignment> {
-
-  specialTypes: Array<String> = ['file', 'questions']
+  specialTypes: Array<string> = ['file', 'questions'];
 
   constructor(override http: HttpClient) {
     super(http);
@@ -30,20 +29,20 @@ export class AssignmentService extends AbstractApiService<Assignment> {
    */
   public override create(instance: Assignment): Observable<Assignment> {
     const form: FormData = new FormData();
-    if (instance._file) form.append('file', instance._file, instance._file.name)
+    if (instance._file) form.append('file', instance._file, instance._file.name);
     if (instance.questions) {
       instance.questions.forEach((question) => {
         if (question._targetCell) {
-          question.targetCell = question._targetCell.fullAddress
+          question.targetCell = question._targetCell.fullAddress;
         }
-        delete question._targetCell
-      })
-      form.append('questions', JSON.stringify(instance.questions))
+        delete question._targetCell;
+      });
+      form.append('questions', JSON.stringify(instance.questions));
     }
-    Object.entries(instance).forEach(([key, value]) => { if (!(key in this.specialTypes) && !(key.charAt(0) == "_")) form.append(key, value) })
-    const obs: Observable<Assignment> = (this.post('assignments/', form) as Observable<Assignment>).pipe(shareReplay(1))
-    obs.subscribe(() => {})
-    return obs
+    Object.entries(instance).forEach(([key, value]) => { if (!(key in this.specialTypes) && !(key.charAt(0) == '_')) form.append(key, value); });
+    const obs: Observable<Assignment> = (this.post('assignments/', form) as Observable<Assignment>).pipe(shareReplay(1));
+    obs.subscribe(() => { });
+    return obs;
   }
 
   /**
@@ -52,7 +51,7 @@ export class AssignmentService extends AbstractApiService<Assignment> {
    * @returns Observable<Assignment>
    */
   public override retrieve(key: string): Observable<Assignment> {
-    return this.get(`assignments/${key}/`) as Observable<Assignment>
+    return this.get(`assignments/${key}/`) as Observable<Assignment>;
   }
 
   /**
@@ -65,16 +64,16 @@ export class AssignmentService extends AbstractApiService<Assignment> {
     if (instance.questions) {
       instance.questions.forEach((question) => {
         if (question._targetCell) {
-          question.targetCell = question._targetCell.fullAddress
+          question.targetCell = question._targetCell.fullAddress;
         }
-        delete question._targetCell
-      })
-      form.append('questions', JSON.stringify(instance.questions))
+        delete question._targetCell;
+      });
+      form.append('questions', JSON.stringify(instance.questions));
     }
-    Object.entries(instance).forEach(([key, value]) => { if (!this.specialTypes.includes(key)) form.append(key, value) })
-    const obs = (this.put(`assignments/${instance.uuid}/`, form) as Observable<Assignment>).pipe(shareReplay(1))
-    obs.subscribe(() => {})
-    return obs
+    Object.entries(instance).forEach(([key, value]) => { if (!this.specialTypes.includes(key)) form.append(key, value); });
+    const obs = (this.put(`assignments/${instance.uuid}/`, form) as Observable<Assignment>).pipe(shareReplay(1));
+    obs.subscribe(() => { });
+    return obs;
   }
 
   /**
@@ -82,9 +81,9 @@ export class AssignmentService extends AbstractApiService<Assignment> {
    * @param instance
    */
   public override destroy(instance: Assignment): Observable<boolean> {
-    const obs = (this.delete(`assignments/${instance.uuid}/`, instance) as Observable<boolean>).pipe(shareReplay(1))
-    obs.subscribe(() => {})
-    return obs
+    const obs = (this.delete(`assignments/${instance.uuid}/`, instance) as Observable<boolean>).pipe(shareReplay(1));
+    obs.subscribe(() => { });
+    return obs;
   }
 
   /**
@@ -93,9 +92,9 @@ export class AssignmentService extends AbstractApiService<Assignment> {
    * @returns Observable<Blob>
    */
   public getFile(assignment: Assignment): Observable<Blob> {
-    if (assignment._file) return of(assignment._file)
-    const obs = (this.get(`files/` + assignment.file.split('/').pop(), {responseType: 'blob'} as Partial<HttpHeaders>) as Observable<Blob>).pipe(shareReplay(1))
-    obs.subscribe((file) => { assignment._file = file });
+    if (assignment._file) return of(assignment._file);
+    const obs = (this.get(`files/${assignment.file.split('/').pop()}`, { responseType: 'blob' } as Partial<HttpHeaders>) as Observable<Blob>).pipe(shareReplay(1));
+    obs.subscribe((file) => { assignment._file = file; });
     return obs;
   }
 }
