@@ -1,35 +1,35 @@
 import { Injectable } from '@angular/core';
-import { Question, QType } from "./question";
-import {Cell} from "exceljs";
-import { WorkbookService } from "../workbook/workbook.service";
-import {Assignment} from "../api/assignment/assignment";
+import { Cell } from 'exceljs';
+import { Question, QType } from './question';
+import { WorkbookService } from '../workbook/workbook.service';
+import { Assignment } from '../api/assignment/assignment';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class QuestionService {
-  questions: Array<Question> = []
+  questions: Array<Question> = [];
+
   constructor(private workbookService: WorkbookService) { }
 
-
   loadQuestions(assignment: Assignment) {
-    this.questions = assignment.questions
+    this.questions = assignment.questions;
   }
 
   deleteQuestions() {
-    this.questions = []
+    this.questions = [];
   }
 
   getQuestions() {
-    return this.questions
+    return this.questions;
   }
 
   addQuestion() {
     const question: Question = {
       type: QType.Value,
-      points: 1
-    }
-    this.questions.push(question)
+      points: 1,
+    };
+    this.questions.push(question);
   }
 
   removeQuestion(question: Question) {
@@ -40,8 +40,8 @@ export class QuestionService {
   }
 
   getTargetCell(question: Question): Cell | undefined {
-    if (question._targetCell) return question._targetCell
-    if (question.targetCell == null) return undefined
-    return this.workbookService.getCell(question.targetCell)
+    if (question._targetCell) return question._targetCell;
+    if (question.targetCell == null) return undefined;
+    return this.workbookService.getCell(question.targetCell);
   }
 }
