@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { AuthService } from "../../../services/api/auth-service/auth.service";
-import { FormBuilder, Validators } from "@angular/forms";
+import { FormBuilder, Validators } from '@angular/forms';
+import { AuthService } from '../../../services/api/auth-service/auth.service';
 
 interface UserCredentials {
   username: string|null
@@ -10,25 +10,23 @@ interface UserCredentials {
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
-  styleUrls: ['./login.component.scss']
+  styleUrls: ['./login.component.scss'],
 })
 export class LoginComponent {
-  logInForm
+  logInForm;
 
   constructor(private formBuilder: FormBuilder, private authService: AuthService) {
     this.logInForm = this.formBuilder.group({
       username: ['', Validators.required],
-      password: ['', Validators.required]
+      password: ['', Validators.required],
     });
   }
 
   onSubmit(formData: Partial<UserCredentials>): void {
     if (this.logInForm.invalid) {
       console.log(this.logInForm.errors);
-    } else {
-      if ('username' in formData && formData.username != null && 'password' in formData && formData.password != null) {
-        this.authService.login(formData.username, formData.password);
-      }
+    } else if ('username' in formData && formData.username != null && 'password' in formData && formData.password != null) {
+      this.authService.login(formData.username, formData.password);
     }
   }
 }
