@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-grader',
   templateUrl: './grader.component.html',
-  styleUrls: ['./grader.component.scss']
+  styleUrls: ['./grader.component.scss'],
 })
 export class GraderComponent {
 
